create shared bottom sheet content component
