+++ messages/en.json
@@ -15,12 +15,6 @@
   "screen.Home.title": {
     "message": "Mapeo Mobile Testing"
   },
-  "screen.sync.Devices.localDeviceDescription": {
-    "message": "These devices are on your project and on the same wifi network as you."
-  },
-  "screen.sync.Devices.localDevices": {
-    "message": "Local Devices"
-  },
   "screen.sync.Devices.searching": {
     "message": "Searching for devices"
   },
@@ -35,6 +29,12 @@
   },
   "screen.sync.DevicesList.sync": {
     "message": "Sync"
+  },
+  "screen.sync.LocalDevicesInfo.localDeviceDescription": {
+    "message": "These devices are on your project and on the same wifi network as you."
+  },
+  "screen.sync.LocalDevicesInfo.localDevices": {
+    "message": "Local Devices"
   },
   "screen.sync.ProjectInfo.connectedNoNetworkName": {
     "message": "Connected (no network name)"
